fix: move transaction titles to event content file
