+++ frontend/src/lib/transactions.ts
@@ -26,37 +26,9 @@ export enum TransactionKeys {
   ETHEREUM_KEY_ROTATE_SUBMISSION = 'ethereumKeyRotateSubmission'
 }
 
-const TRANSACTION_TITLES: Record<TransactionKeys, string> = {
-  [TransactionKeys.UNKNOWN]: 'Unknown transaction',
-  [TransactionKeys.ORDER_SUBMISSION]: 'Order submission',
-  [TransactionKeys.ORDER_CANCELLATION]: 'Order cancellation',
-  [TransactionKeys.ORDER_AMENDMENT]: 'Order amendment',
-  [TransactionKeys.VOTE_SUBMISSION]: 'Vote submission',
-  [TransactionKeys.WITHDRAW_SUBMISSION]: 'Withdraw submission',
-  [TransactionKeys.LIQUIDTY_PROVISION_SUBMISSION]: 'Liquidity provision',
-  [TransactionKeys.LIQUIDTY_PROVISION_CANCELLATION]:
-    'Liquidity provision cancellation',
-  [TransactionKeys.LIQUIDITY_PROVISION_AMENDMENT]:
-    'Liquidity provision amendment',
-  [TransactionKeys.PROPOSAL_SUBMISSION]: 'Proposal submission',
-  [TransactionKeys.ANNOUNCE_NODE]: 'Announce node',
-  [TransactionKeys.NODE_VOTE]: 'Node vote',
-  [TransactionKeys.NODE_SIGNATURE]: 'Node signature',
-  [TransactionKeys.CHAIN_EVENT]: 'Chain event',
-  [TransactionKeys.ORACLE_DATA_SUBMISSION]: 'Oracle data submission',
-  [TransactionKeys.UNDELEGATE_SUBMISSION]: 'Undelegate submission',
-  [TransactionKeys.DELEGATE_SUBMISSION]: 'Delegate submission',
-  [TransactionKeys.TRANSFER]: 'Transfer',
-  [TransactionKeys.CANCEL_TRANSFER]: 'Cancel transfer',
-  [TransactionKeys.KEY_ROTATE_SUBMISSION]: 'Key rotation submission',
-  [TransactionKeys.ETHEREUM_KEY_ROTATE_SUBMISSION]:
-    'Ethereum key rotation submission'
-}
-
 export type Transaction = RequestTransactionSendingContent & {
   type: TransactionKeys
   payload: object
-  title: string
 }
 
 const getPayload = (transaction: string): object => {
@@ -78,7 +50,6 @@ export const parseTransaction = (data: RequestTransactionSendingContent) => {
   return {
     ...data,
     type,
-    payload,
-    title: TRANSACTION_TITLES[type]
+    payload
   }
 }
